Version 3 - Split Functional, Revised main.py
Split is now fully functional and fully tested for interactive players. main.py is revamped so that there is a play_game option that asks user for input for how many interactive users will play. This differentiates run_sim() which will be for testing strategies.
To do: test split() for bots, write run_sim() fct. Maybe add a master function that will ask for mode -interactive or sim. Or should it be different files.